# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World Cup Tickets API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Match {
  id: number;
  homeTeam: string;
  awayTeam: string;
  homeTeamAr: string;
  awayTeamAr: string;
  /** @nullable */
  homeTeamFlag?: string | null;
  /** @nullable */
  awayTeamFlag?: string | null;
  stadium: string;
  stadiumAr: string;
  city: string;
  cityAr: string;
  matchDate: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  descriptionAr?: string | null;
  stage: string;
  stageAr: string;
  minPrice: number;
  isActive: boolean;
  createdAt: string;
}

export interface MatchInput {
  homeTeam: string;
  awayTeam: string;
  homeTeamAr: string;
  awayTeamAr: string;
  homeTeamFlag?: string;
  awayTeamFlag?: string;
  stadium: string;
  stadiumAr: string;
  city: string;
  cityAr: string;
  matchDate: string;
  image?: string;
  description?: string;
  descriptionAr?: string;
  stage: string;
  stageAr: string;
  minPrice: number;
  isActive?: boolean;
}

export interface MatchUpdate {
  homeTeam?: string;
  awayTeam?: string;
  homeTeamAr?: string;
  awayTeamAr?: string;
  homeTeamFlag?: string;
  awayTeamFlag?: string;
  stadium?: string;
  stadiumAr?: string;
  city?: string;
  cityAr?: string;
  matchDate?: string;
  image?: string;
  description?: string;
  descriptionAr?: string;
  stage?: string;
  stageAr?: string;
  minPrice?: number;
  isActive?: boolean;
}

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  available: "available",
  reserved: "reserved",
  sold: "sold",
} as const;

export interface Ticket {
  id: number;
  matchId: number;
  category: string;
  categoryAr: string;
  seatNumber: string;
  section: string;
  /** @nullable */
  row?: string | null;
  price: number;
  status: TicketStatus;
  createdAt: string;
}

export interface TicketInput {
  matchId: number;
  category: string;
  categoryAr: string;
  seatNumber: string;
  section: string;
  row?: string;
  price: number;
}

export type TicketUpdateStatus =
  (typeof TicketUpdateStatus)[keyof typeof TicketUpdateStatus];

export const TicketUpdateStatus = {
  available: "available",
  reserved: "reserved",
  sold: "sold",
} as const;

export interface TicketUpdate {
  category?: string;
  categoryAr?: string;
  seatNumber?: string;
  section?: string;
  row?: string;
  price?: number;
  status?: TicketUpdateStatus;
}

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  email: string;
  country: string;
  /** @nullable */
  notes?: string | null;
  matchId: number;
  /** @nullable */
  matchName?: string | null;
  selectedSeats: string;
  totalPrice: number;
  paymentStatus: OrderPaymentStatus;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  phone: string;
  email: string;
  country: string;
  notes?: string;
  matchId: number;
  selectedSeats: string;
  totalPrice: number;
}

export type OrderUpdatePaymentStatus =
  (typeof OrderUpdatePaymentStatus)[keyof typeof OrderUpdatePaymentStatus];

export const OrderUpdatePaymentStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;

export interface OrderUpdate {
  paymentStatus?: OrderUpdatePaymentStatus;
}

export interface Post {
  id: number;
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  /** @nullable */
  image?: string | null;
  isPublished: boolean;
  createdAt: string;
}

export interface PostInput {
  title: string;
  titleAr: string;
  content: string;
  contentAr: string;
  image?: string;
  isPublished?: boolean;
}

export interface PostUpdate {
  title?: string;
  titleAr?: string;
  content?: string;
  contentAr?: string;
  image?: string;
  isPublished?: boolean;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminAuthResult {
  token: string;
  message: string;
}

export interface StatsSummary {
  totalMatches: number;
  totalOrders: number;
  pendingOrders: number;
  confirmedOrders: number;
  cancelledOrders: number;
  totalRevenue: number;
  availableTickets: number;
  totalPosts: number;
}

export type ListMatchesParams = {
  search?: string;
  sortBy?: ListMatchesSortBy;
  sortOrder?: ListMatchesSortOrder;
};

export type ListMatchesSortBy =
  (typeof ListMatchesSortBy)[keyof typeof ListMatchesSortBy];

export const ListMatchesSortBy = {
  date: "date",
  price: "price",
} as const;

export type ListMatchesSortOrder =
  (typeof ListMatchesSortOrder)[keyof typeof ListMatchesSortOrder];

export const ListMatchesSortOrder = {
  asc: "asc",
  desc: "desc",
} as const;

export type ListOrdersParams = {
  status?: ListOrdersStatus;
};

export type ListOrdersStatus =
  (typeof ListOrdersStatus)[keyof typeof ListOrdersStatus];

export const ListOrdersStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
} as const;