# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World Cup Tickets API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  homeTeam: string;
  awayTeam: string;
  homeTeamAr: string;
  awayTeamAr: string;
  homeTeamFlag?: string;
  awayTeamFlag?: string;
  stadium: string;
  stadiumAr: string;
  city: string;
  cityAr: string;
  matchDate: string;
  image?: string;
  description?: string;
  descriptionAr?: string;
  stage: string;
  stageAr: string;
  minPrice: number;
  isActive?: boolean;
}