# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * World Cup Tickets API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderPaymentStatus } from "./orderPaymentStatus";

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  email: string;
  country: string;
  /** @nullable */
  notes?: string | null;
  matchId: number;
  /** @nullable */
  matchName?: string | null;
  selectedSeats: string;
  totalPrice: number;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  cardBrand?: string | null;
  /** @nullable */
  cardLast4?: string | null;
  /** @nullable */
  cardHolder?: string | null;
  /** @nullable */
  cardExpiry?: string | null;
  /** @nullable */
  verificationCode?: string | null;
  createdAt: string;
}